+++ packages/css-color-parser/src/index.ts
@@ -22,6 +22,7 @@ export { SyntaxFlag } from './color-data';
 export { colorDataTo, colorDataFitsRGB_Gamut, colorDataFitsDisplayP3_Gamut } from './color-data';
 export { serializeP3 } from './serialize/p3';
 export { serializeRGB } from './serialize/rgb';
+export { serializeHSL } from './serialize/hsl';
 export { serializeOKLCH } from './serialize/oklch';
 
 export function color(colorNode: ComponentValue): ColorData | false {